Alert

+++ components/FavoritesComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
@@ -25,7 +25,7 @@ class Favorites extends Component {
     }
 
     render() {
-        
+
         const { navigate } = this.props.navigation;
 
         const renderFavoriteItem = ({ item }) => {
@@ -33,7 +33,26 @@ class Favorites extends Component {
             {
               text: "Delete",
               type: "delete",
-              onPress: () => this.props.deleteFavorite(item.id),
+              onPress: () => {
+                Alert.alert(
+                  "Delete Favorite?",
+                  "Are you sure you wish to delete the favorite campsite " +
+                    item.name +
+                    "?",
+                  [
+                    {
+                      text: "Cancel",
+                      onPress: () => console.log(item.name + "Not Deleted"),
+                      style: " cancel",
+                    },
+                    {
+                      text: "OK",
+                      onPress: () => this.props.deleteFavorite(item.id),
+                    },
+                  ],
+                  { cancelable: false }
+                );
+              },
             },
           ];
